Revised Classes
I made a package system since everyone's packages were all over the place. I also went ahead and touched up each of the classes and left comments and documentation on them. As of now, the only things left to finish are all of the expenses in ExpenseHandler and the GameManager class.