Adding NewsScreen and making adjustments to other classes
I finished a draft of NewsScreen that should work. If you want to test it, see the other branch that has it set up to test the NewsScreen.

Player was missing a starting balance assignment, so I added one in the constructor.

In GameManager, I filled out startGame(), getCost(), goToNewsScreen(), goToNextWeek(), generateBusinessWeeklyUpdate(), and added an initializing method for EventHandler and ExpenseHandler.