Fixed issue with bankLabel
On the stock screen where it shows the player's name and bank account, it doesn't correctly add the " 's " to the name. I also noticed that there was an unused public method in Stock Screen, so I deleted the if statement in init() and used the method in Stock Screen. I also changed that method from public to private.